moving BossStats to own component in prep for adding routing
